refactor(ast): extract import/export/source merge into fileMerge.ts (#3572 1f)
Move combineImports, combineExports, and combineSources (plus the sourceKey,
pathTypeKey, exportKey, importKey, sortKey, and mergeNameArrays helpers) out of
utils/ast.ts into utils/fileMerge.ts, with their tests in fileMerge.test.ts.
factory.ts now imports the combine helpers from fileMerge.ts. The @kubb/ast and
@kubb/ast/factory barrels are unchanged.
